Reutilice la componente de agregar cliente para editar y realice validaciones en el servicio, elimine la componente de editr-cliente

+++ src/app/services/clientes.service.ts
@@ -57,16 +57,18 @@ export class ClienteServices {
   }
 
   /**---Modificamos la información del cliente--- */
-  modificarCliente(cliente: Cliente){
-     this.clienteDoc = this.firebaseDb.doc(`clientes/${cliente.id}`);
+  modificarCliente(cliente: Cliente, id: string){
+     this.clienteDoc = this.firebaseDb.doc(`clientes/${id}`);
 
      this.clienteDoc.update(cliente);
   }
 
   /**---Elimina la información del cliente--- */
   eliminarCliente(cliente: Cliente){
-    this.clienteDoc = this.firebaseDb.doc(`clientes/${cliente.id}`);
-
-    this.clienteDoc.delete();
+    if(cliente){
+      this.clienteDoc = this.firebaseDb.doc(`clientes/${cliente.id}`);
+  
+      this.clienteDoc.delete();
+    }
  }
 }
